shakerPart in react native
