+++ packages/frontend/mobile/src/components/form-cocktail/LevelPart.tsx
@@ -7,41 +7,40 @@ import colors from '../../../colors';
 
 const numberLevel = [3, 2, 1];
 
+const styles = StyleSheet.create({
+  title: {
+    color: colors.dark,
+    fontSize: 24,
+    fontWeight: 'bold',
+    marginBottom: 15,
+  },
+  container: {
+    flexDirection: 'row-reverse',
+    position: 'relative',
+    bottom: '3%',
+    left: '-4%',
+    marginTop: 15,
+  },
+  levelContainer: {
+    height: 90,
+    width: 70,
+    backgroundColor: 'transparent',
+    justifyContent: 'center',
+    alignItems: 'center',
+    marginHorizontal: 5,
+    marginBottom: 80,
+  },
+  image: {
+    height: 90,
+    width: 70,
+    resizeMode: 'cover',
+  },
+});
 export default function LevelPart({
   level,
   handleLevelClick,
   errors,
 }: LevelPartProps) {
-  const styles = StyleSheet.create({
-    title: {
-      color: colors.dark,
-      fontSize: 24,
-      fontWeight: 'bold',
-      marginBottom: 15,
-    },
-    container: {
-      flexDirection: 'row-reverse',
-      position: 'relative',
-      bottom: '3%',
-      left: '-4%',
-      marginTop: 15,
-    },
-    levelContainer: {
-      height: 90,
-      width: 70,
-      backgroundColor: 'transparent',
-      justifyContent: 'center',
-      alignItems: 'center',
-      marginHorizontal: 5,
-      marginBottom: 80,
-    },
-    image: {
-      height: 90,
-      width: 70,
-      resizeMode: 'cover',
-    },
-  });
-
   return (
     <>
       <Text style={styles.title}>{'CHOOSE YOUR DUEL'}</Text>
